refactor: Judge domain으로 이동 및 Dealer와 Player를 속성으로 가지게 변경
